requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
whisper-openai
scikit-learn
numpy==1.26.4
scipy
transformers
vaderSentiment
pytest
httpx